pyproject.toml
[tool.poetry.dependencies]
python        = "^3.10"
pandas        = "^2.0"
requests      = "^2.31"
psycopg2-binary = "^2.9"
sqlalchemy    = "^2.0"
python-dotenv = "^1.0"
matplotlib    = "^3.8"
seaborn       = "^0.13"